Docker/requirements.txt
flask
absl-py==2.1.0
aiobotocore==2.7.0
aiohttp==3.9.3
aioitertools==0.7.1
aiosignal==1.2.0
alabaster==0.7.12
altair==5.0.1
annotated-types==0.6.0
anyio==4.3.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
astroid==2.14.2
astropy==5.3.4
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
atomicwrites==1.4.0
attrs==23.2.0
Automat==20.2.0
autopep8==2.0.4
Babel==2.14.0
backports.functools-lru-cache==1.6.4
backports.tempfile==1.0
backports.weakref==1.0.post1
bcrypt==3.2.0
beautifulsoup4==4.12.3
binaryornot==0.4.4
black==24.3.0
bleach==6.1.0
blinker==1.6.2
bokeh==3.4.0
boltons==23.0.0
botocore==1.31.64
Bottleneck==1.3.7
Brotli==1.0.9
cachetools==4.2.2
cairocffi==1.6.1
CairoSVG==2.7.1
certifi==2024.2.2
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==2.2.1
colorama==0.4.6
colorcet==3.1.0
comm==0.2.2
constantly==23.10.4
contourpy==1.2.1
cookiecutter==2.6.0
cryptography==41.0.3
cssselect==1.2.0
cssselect2==0.7.0
cycler==0.12.1
cytoolz==0.12.2
dask==2023.11.0
datashader==0.16.0
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20200713
dill==0.3.7
distlib==0.3.8
distributed==2023.11.0
docstring-to-markdown==0.11
docutils==0.18.1
entrypoints==0.4
et-xmlfile==1.1.0
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.13.1
flake8==7.0.0
Flask==2.2.5
flatbuffers==24.3.25
fonttools==4.51.0
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.10.0
future==0.18.3
gast==0.5.4
gensim==4.3.0
gitdb==4.0.7
GitPython==3.1.37
gmpy2==2.1.2
google-auth==2.6.0
google-auth-oauthlib==0.5.2
google-pasta==0.2.0
greenlet==3.0.1
grpcio==1.62.2
h11==0.14.0
h5py==3.11.0
HeapDict==1.0.1
holoviews==1.18.3
httpcore==1.0.5
httpx==0.27.0
hvplot==0.9.2
hyperlink==21.0.0
idna==3.4
imagecodecs==2023.1.23
imageio==2.33.1
imagesize==1.4.1
imbalanced-learn==0.11.0
importlib-metadata==7.0.1
incremental==22.10.0
inflection==0.5.1
iniconfig==1.1.1
intake==0.6.8
intervaltree==3.1.0
ipykernel==6.29.4
ipython==8.23.0
ipython-genutils==0.2.0
ipywidgets==8.1.2
isoduration==20.11.0
isort==5.9.3
itemadapter==0.3.0
itemloaders==1.1.0
itsdangerous==2.0.1
jaraco.classes==3.2.1
jedi==0.19.1
jeepney==0.7.1
jellyfish==1.0.1
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.2.0
json5==0.9.25
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter_client==8.6.1
jupyter-console==6.6.3
jupyter_core==5.7.2
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_server==2.14.0
jupyter_server_terminals==0.5.3
jupyterlab==4.1.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.26.0
jupyterlab_widgets==3.0.10
keras==3.2.1
Keras-Preprocessing==1.1.2
keyring==24.3.1
kiwisolver==1.4.5
lazy_loader==0.3
lazy-object-proxy==1.10.0
lckr_jupyterlab_variableinspector==3.1.0
libarchive-c==2.9
libclang==18.1.1
linkify-it-py==2.0.0
llvmlite==0.42.0
lmdb==1.4.1
locket==1.0.0
lxml==4.9.3
lz4==4.3.2
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.7
mccabe==0.7.0
mdit-py-plugins==0.3.0
mdurl==0.1.2
mistune==3.0.2
mkl-fft==1.3.8
mkl-random==1.2.4
mkl-service==2.4.0
ml-dtypes==0.3.2
more-itertools==10.1.0
mpmath==1.3.0
msgpack==1.0.3
multidict==6.0.4
multipledispatch==0.6.0
mypy==1.8.0
mypy-extensions==1.0.0
mysql-connector-python
namex==0.0.8
navigator-updater
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.1
nltk==3.8.1
notebook==7.1.3
notebook_shim==0.2.4
numba==0.59.1
numexpr==2.8.7
numpy==1.26.4
numpydoc==1.5.0
oauthlib==3.2.2
opencv-python==4.9.0.80
openpyxl==3.0.10
opt-einsum==3.3.0
optree==0.11.0
overrides==7.7.0
packaging==23.2
pandas==2.2.2
pandocfilters==1.5.1
panel==1.4.1
param==2.1.0
parsel==1.8.1
parso==0.8.4
partd==1.4.1
pathspec==0.10.3
patsy==0.5.3
pexpect==4.8.0
pickleshare==0.7.5
pillow==10.3.0
pip==24.0
pkce==1.0.3
pkginfo==1.9.6
platformdirs==4.2.0
plotly==5.19.0
pluggy==1.0.0
ply==3.11
prometheus_client==0.20.0
prompt-toolkit==3.0.43
Protego==0.1.16
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
py4j==0.10.9.7
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.11.1
pycosat==0.6.6
pycparser==2.22
pyct==0.5.0
pycurl==7.45.2
pydantic==2.5.3
pydantic_core==2.14.6
pydeck==0.8.0
PyDispatcher==2.0.5
pydocstyle==6.3.0
pyerfa==2.0.0
pyflakes==3.2.0
Pygments==2.17.2
PyJWT==2.4.0
pylint==2.16.2
pylint-venv==3.0.3
pyls-spyder==0.4.0
mysql-connector-python==8.0.5
pyodbc==5.0.1
pyOpenSSL==23.2.0
pyparsing==3.0.9
PyQt5==5.15.10
PyQt5-sip==12.13.0
PyQtWebEngine==5.15.6
PySocks==1.7.1
pyspark==3.5.1
pytest==7.4.0
python-dateutil==2.9.0.post0
python-dotenv==0.21.0
python-json-logger==2.0.7
python-lsp-black==2.0.0
python-lsp-jsonrpc==1.1.2
python-lsp-server==1.10.0
python-slugify==5.0.2
python-snappy==0.6.1
pytoolconfig==1.2.6
pytz==2023.3.post1
pyviz_comms==3.0.2
pywavelets==1.5.0
pyxdg==0.27
PyYAML==6.0.1
pyzbar==0.1.9
pyzmq==26.0.2
QDarkStyle==3.2.3
qstylizer==0.2.2
QtAwesome==1.2.2
qtconsole==5.5.1
QtPy==2.4.1
queuelib==1.6.2
referencing==0.34.0
regex==2023.10.3
requests==2.31.0
requests-file==1.5.1
requests-oauthlib==1.3.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rope==1.12.0
rpds-py==0.18.0
rsa==4.7.2
Rtree==1.0.1
s3fs==2023.10.0
scikit-image==0.22.0
scikit-learn==1.3.0
scipy==1.12.0
Scrapy==2.11.1
seaborn==0.12.2
SecretStorage==3.3.1
semver==3.0.2
Send2Trash==1.8.3
service-identity==18.1.0
setuptools==68.2.2
sip==6.7.12
six==1.16.0
smart-open==5.2.1
smmap==4.0.0
sniffio==1.3.1
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.5
Sphinx==5.0.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
spyder==5.5.1
spyder-kernels==2.5.0
SQLAlchemy==2.0.25
st==0.0.8
stack-data==0.6.3
statsmodels==0.14.0
streamlit==1.33.0
sympy==1.12
tables==3.9.2
tabulate==0.9.0
tblib==1.7.0
tenacity==8.2.2
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.6.0
tensorflow==2.16.1
tensorflow-estimator==2.12.0
tensorflow-hub==0.16.1
tensorflow-io-gcs-filesystem==0.36.0
termcolor==2.4.0
terminado==0.18.1
text-unidecode==1.3
textdistance==4.2.1
tf_keras==2.16.0
threadpoolctl==2.2.0
three-merge==0.1.1
tifffile==2023.4.12
tinycss2==1.2.1
tldextract==3.2.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.1
toolz==0.12.0
tornado==6.4
tqdm==4.65.0
traitlets==5.14.3
truststore==0.8.0
Twisted==23.10.0
types-python-dateutil==2.9.0.20240316
typing_extensions==4.11.0
tzdata==2024.1
uc-micro-py==1.0.1
ujson==5.4.0
unicodedata2==15.1.0
Unidecode==1.2.0
uri-template==1.3.0
urllib3==2.0.7
virtualenv==20.26.0
w3lib==2.1.2
watchdog==2.1.6
wcwidth==0.2.5
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.0.2
whatthepatch==1.0.2
wheel==0.41.2
widgetsnbextension==4.0.10
wrapt==1.16.0
wurlitzer==3.0.2
xarray==2023.6.0
xyzservices==2022.9.0
yapf==0.40.2
yarl==1.9.3
zict==3.0.0
zipp==3.17.0
zope.interface==5.4.0
zstandard==0.19.0
